Create discovery screen in feedResources
