+++ feedResources/app/src/main/AndroidManifest.xml
@@ -8,16 +8,20 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.FireworkExample"
-        tools:targetApi="31">
+        android:theme="@style/Theme.FireworkExample">
+
         <activity
             android:name="tv.fw.example.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name="tv.fw.example.discovery.DiscoveryActivity" />
+
     </application>
 
 </manifest>
