+++ feedResources/app/src/main/java/tv/fw/example/MainActivity.kt
@@ -1,42 +1,18 @@
 package tv.fw.example
 
-import android.annotation.SuppressLint
 import android.os.Bundle
-import android.view.LayoutInflater
 import androidx.appcompat.app.AppCompatActivity
-import tv.fw.common.feed.FeedResource
+import tv.fw.example.discovery.DiscoveryActivity
 import tv.fw.example.feedResources.databinding.ActivityMainBinding
-import tv.fw.videofeed.options.ViewOptions
 
 class MainActivity : AppCompatActivity() {
-
     private lateinit var binding: ActivityMainBinding
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(LayoutInflater.from(this))
+        binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        setupDetails()
-
-        initVideoFeedView()
-    }
-
-    @SuppressLint("SetTextI18n")
-    private fun setupDetails() {
-        binding.details.source.text = "Discovery"
-        binding.details.channel.text = "N/A"
-        binding.details.playlist.text = "N/A"
-    }
-
-    private fun initVideoFeedView() {
-        val videoFeedView = binding.videoFeedView
-
-        val viewOptions = ViewOptions.Builder()
-            // Check Channel, Playlist, or Dynamic example apps for other feed sources
-            .feedResource(FeedResource.Discovery)
-            .build()
-
-        videoFeedView.init(viewOptions)
+        binding.btnDiscovery.setOnClickListener {
+            startActivity(DiscoveryActivity.intent(this))
+        }
     }
 }
