+++ feedResources/app/src/main/java/tv/fw/example/discovery/DiscoveryActivity.kt
@@ -1,0 +1,51 @@
+package tv.fw.example.discovery
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.view.LayoutInflater
+import androidx.appcompat.app.AppCompatActivity
+import tv.fw.common.feed.FeedResource
+import tv.fw.example.feedResources.databinding.ActivityDiscoveryBinding
+import tv.fw.videofeed.baseOptions
+import tv.fw.videofeed.viewOptions
+
+class DiscoveryActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityDiscoveryBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityDiscoveryBinding.inflate(LayoutInflater.from(this))
+        setContentView(binding.root)
+
+        setupDetails()
+
+        initVideoFeedView()
+    }
+
+    @SuppressLint("SetTextI18n")
+    private fun setupDetails() {
+        binding.details.source.text = "Discovery"
+        binding.details.channel.text = "N/A"
+        binding.details.playlist.text = "N/A"
+    }
+
+    private fun initVideoFeedView() {
+        val videoFeedView = binding.fwVideoFeedView
+
+        val viewOptions = viewOptions {
+            baseOptions {
+                // Check Channel, Playlist, or Dynamic screens for other feed sources
+                feedResource(FeedResource.Discovery)
+            }
+        }
+
+        videoFeedView.init(viewOptions)
+    }
+
+    companion object {
+        fun intent(context: Context) = Intent(context, DiscoveryActivity::class.java)
+    }
+}
