+++ feedResources/app/src/main/res/layout/activity_main.xml
@@ -1,23 +1,47 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context="tv.fw.example.MainActivity">
 
-    <tv.fw.videofeed.VideoFeedView
-        android:id="@+id/videoFeedView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@id/details"
+
+    <Button
+        android:id="@+id/btn_discovery"
+        android:layout_width="200dp"
+        android:layout_height="75dp"
+        android:layout_marginTop="16dp"
+        android:text="@string/discovery"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <include
-        android:id="@+id/details"
-        layout="@layout/details"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent" />
+    <Button
+        android:id="@+id/btn_channel"
+        android:layout_width="200dp"
+        android:layout_height="75dp"
+        android:text="@string/channel"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btn_discovery" />
+
+    <Button
+        android:id="@+id/btn_playlist"
+        android:layout_width="200dp"
+        android:layout_height="75dp"
+        android:text="@string/playlist"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btn_channel" />
+
+    <Button
+        android:id="@+id/btn_dynamic_content"
+        android:layout_width="200dp"
+        android:layout_height="75dp"
+        android:text="@string/dynamic_content"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btn_playlist" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
